Patchwerked edition of classic quest text
I modified the classic quest text addon to work in the current 100200 version of retail wow.

It works as long as you do not touch the config menu (which does not work half the time), if you do the an empty box will appear in the quest window that will cause the quest rewards to be hidden, you can clear this by using the curseforge client to delete the addon variables, or manually removing the relevant variables files.